+++ include/odp/arch/arm32-linux/odp/api/abi/event.h
@@ -4,4 +4,4 @@
  * SPDX-License-Identifier:     BSD-3-Clause
  */
 
-#include <odp/arch/default/api/abi/event.h>
+#include <odp/api/abi-default/event.h>
