+++ include/odp/arch/x86_32-linux/odp/api/abi/pool.h
@@ -4,4 +4,4 @@
  * SPDX-License-Identifier:     BSD-3-Clause
  */
 
-#include <odp/arch/default/api/abi/pool.h>
+#include <odp/api/abi-default/pool.h>
